[YITEAM-56] 회원가입 비밀번호 입력 화면 구현 - JoinViewController
